+++ src/main/java/com/example/ceresto/eat/enumerati/AllergeniciEnum.java
@@ -1,22 +1,22 @@
 package com.example.ceresto.eat.enumerati;
 
 public enum AllergeniciEnum {
-    LATTOSIO(1, "Latte", "This courses countains milk and derivatives"),
-    GLUTINE(2, "Grano", "This courses countains gluten and derivatives"),
-    ARACHIDI(3, "Arachidi", "This courses countains arach and derivatives"),
-    FRUTTA_A_GUSCIO(4, "Frutta a guscio", "This courses countains Walnuts and derivatives nuts"),
-    SOIA(5, "Soia", "This courses countains Soia and derivatives"),
-    UOVA(6, "Uova", "This courses countains eggs and derivatives"),
-    PESCE(7, "Pesce", "This courses countains fish and derivatives"),
-    CROSTACEI(8, "Crostacei", "This courses countains shellfish and derivatives"),
-    SENAPE(9, "Senape", "This courses countains mustard and derivatives"),
-    SESAMO(10, "Sesamo", "This courses countains sesame and derivatives"),
-    SOLFITI(11, "Solfiti", "This courses countains sulfites and derivatives");
+    LATTOSIO(1L, "Latte", "This courses countains milk and derivatives"),
+    GLUTINE(2L, "Grano", "This courses countains gluten and derivatives"),
+    ARACHIDI(3L, "Arachidi", "This courses countains arach and derivatives"),
+    FRUTTA_A_GUSCIO(4L, "Frutta a guscio", "This courses countains Walnuts and derivatives nuts"),
+    SOIA(5L, "Soia", "This courses countains Soia and derivatives"),
+    UOVA(6L, "Uova", "This courses countains eggs and derivatives"),
+    PESCE(7L, "Pesce", "This courses countains fish and derivatives"),
+    CROSTACEI(8L, "Crostacei", "This courses countains shellfish and derivatives"),
+    SENAPE(9L, "Senape", "This courses countains mustard and derivatives"),
+    SESAMO(10L, "Sesamo", "This courses countains sesame and derivatives"),
+    SOLFITI(11L, "Solfiti", "This courses countains sulfites and derivatives");
     private String name;
     private String description;
-    private Integer id;
+    private Long id;
     
-    AllergeniciEnum(Integer id, String name, String description) {
+    AllergeniciEnum(Long id, String name, String description) {
         this.name = name;
         this.id = id;
         this.description = description;
@@ -28,10 +28,10 @@ public enum AllergeniciEnum {
     public void setName(String name) {
         this.name = name;
     }
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
-    public void setId(int id) {
+    public void setId(Long id) {
         this.id = id;
     }
